# src/config/swagger.yaml
openapi: 3.0.3
info:
  title: Graph Optimization API
  version: 1.0.0
  description: |
    API for managing graph optimization models using Dijkstra algorithm.
    Built with Node.js, Express, TypeScript, Sequelize, PostgreSQL.
    Authentication with JWT RS256.
    Token costs: model creation and execution = `0.20 * nodes + 0.01 * edges`.
servers:
  - url: http://localhost:3000
    description: Local development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, example: user02@test.com }
                password: { type: string, example: "12345678" }
      responses:
        "201":
          description: User registered
          content:
            application/json:
              example:
                message: User registered successfully
                id: 2
                email: user02@test.com
        "400":
          description: Invalid input
          content:
            application/json:
              example: { error: "Invalid email or password" }
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Authenticate and obtain JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, example: user02@test.com }
                password: { type: string, example: "12345678" }
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              example:
                token: eyJhbGciOi...
                expiresIn: 3600
        "401":
          description: Unauthorized
          content:
            application/json:
              example: { error: "Invalid credentials" }
...